@@ -14,10 +14,12 @@ version = '1.0'
 group = 'com.yourname.modid'
 archivesBaseName = 'modid'
 
-java.toolchain.languageVersion = JavaLanguageVersion.of(8)
+java.toolchain.languageVersion = JavaLanguageVersion.of(17)
 
 minecraft {
-    mappings channel: 'snapshot', version: '20201028-1.16.3'
+    mappings channel: 'official', version: '1.18.2'
+
+    accessTransformer = file('src/main/resources/META-INF/accesstransformer.cfg')
 
     runs {
         client {
@@ -25,11 +27,6 @@ minecraft {
 
             property 'forge.logging.markers', 'REGISTRIES'
             property 'forge.logging.console.level', 'debug'
-
-            jvmArgs = [
-                    "--add-exports=java.base/sun.security.util=ALL-UNNAMED",
-                    "--add-opens=java.base/java.util.jar=ALL-UNNAMED"
-            ]
 
             mods {
                 examplemod {
@@ -44,11 +41,6 @@ minecraft {
             property 'forge.logging.markers', 'REGISTRIES'
             property 'forge.logging.console.level', 'debug'
 
-            jvmArgs = [
-                    "--add-exports=java.base/sun.security.util=ALL-UNNAMED",
-                    "--add-opens=java.base/java.util.jar=ALL-UNNAMED"
-            ]
-
             mods {
                 examplemod {
                     source sourceSets.main
@@ -59,7 +51,11 @@ minecraft {
 }
 
 dependencies {
-    minecraft 'net.minecraftforge:forge:1.16.5-36.2.26'
+    minecraft 'net.minecraftforge:forge:1.18.2-40.1.0'
+}
+
+tasks.withType(JavaCompile).configureEach {
+    options.encoding = 'UTF-8' // Use the UTF-8 charset for Java compilation
 }
 
 apply from: 'mcreator.gradle'
