@@ -1,15 +1,11 @@
 /*
  *    MCreator note:
  *
- *    If you lock base mod element files, you can edit this file and the proxy files
- *    and they won't get overwritten. If you change your mod package or modid, you
- *    need to apply these changes to this file MANUALLY.
+ *    If you lock base mod element files, you can edit this file and it won't get overwritten.
+ *    If you change your modid or package, you need to apply these changes to this file MANUALLY.
  *
  *    Settings in @Mod annotation WON'T be changed in case of the base mod element
  *    files lock too, so you need to set them manually here in such case.
- *
- *    Keep the FundysCursedFoodModElements object in this class and all calls to this object
- *    INTACT in order to preserve functionality of mod elements generated by MCreator.
  *
  *    If you do not lock base mod element files in Workspace settings, this file
  *    will be REGENERATED on each build.
@@ -20,85 +16,43 @@ package cz.pisekpiskovec.fundyscursedfood;
 import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.LogManager;
 
-import net.minecraftforge.fml.network.simple.SimpleChannel;
-import net.minecraftforge.fml.network.NetworkRegistry;
+import net.minecraftforge.network.simple.SimpleChannel;
+import net.minecraftforge.network.NetworkRegistry;
+import net.minecraftforge.network.NetworkEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
-import net.minecraftforge.fml.event.server.FMLServerStartingEvent;
-import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
-import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.event.RegistryEvent;
-import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.eventbus.api.IEventBus;
 
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.item.Item;
-import net.minecraft.entity.EntityType;
-import net.minecraft.enchantment.Enchantment;
-import net.minecraft.block.Block;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.network.FriendlyByteBuf;
 
 import java.util.function.Supplier;
+import java.util.function.Function;
+import java.util.function.BiConsumer;
+
+import cz.pisekpiskovec.fundyscursedfood.init.FundysCursedFoodModTabs;
+import cz.pisekpiskovec.fundyscursedfood.init.FundysCursedFoodModItems;
 
 @Mod("fundys_cursed_food")
 public class FundysCursedFoodMod {
 	public static final Logger LOGGER = LogManager.getLogger(FundysCursedFoodMod.class);
+	public static final String MODID = "fundys_cursed_food";
 	private static final String PROTOCOL_VERSION = "1";
-	public static final SimpleChannel PACKET_HANDLER = NetworkRegistry.newSimpleChannel(
-			new ResourceLocation("fundys_cursed_food", "fundys_cursed_food"), () -> PROTOCOL_VERSION, PROTOCOL_VERSION::equals,
-			PROTOCOL_VERSION::equals);
-	public FundysCursedFoodModElements elements;
+	public static final SimpleChannel PACKET_HANDLER = NetworkRegistry.newSimpleChannel(new ResourceLocation(MODID, MODID), () -> PROTOCOL_VERSION,
+			PROTOCOL_VERSION::equals, PROTOCOL_VERSION::equals);
+	private static int messageID = 0;
 
 	public FundysCursedFoodMod() {
-		elements = new FundysCursedFoodModElements();
-		FMLJavaModLoadingContext.get().getModEventBus().register(this);
-		FMLJavaModLoadingContext.get().getModEventBus().addListener(this::init);
-		FMLJavaModLoadingContext.get().getModEventBus().addListener(this::clientLoad);
-		MinecraftForge.EVENT_BUS.register(new FundysCursedFoodModFMLBusEvents(this));
+		FundysCursedFoodModTabs.load();
+		IEventBus bus = FMLJavaModLoadingContext.get().getModEventBus();
+
+		FundysCursedFoodModItems.REGISTRY.register(bus);
+
 	}
 
-	private void init(FMLCommonSetupEvent event) {
-		elements.getElements().forEach(element -> element.init(event));
-	}
-
-	public void clientLoad(FMLClientSetupEvent event) {
-		elements.getElements().forEach(element -> element.clientLoad(event));
-	}
-
-	@SubscribeEvent
-	public void registerBlocks(RegistryEvent.Register<Block> event) {
-		event.getRegistry().registerAll(elements.getBlocks().stream().map(Supplier::get).toArray(Block[]::new));
-	}
-
-	@SubscribeEvent
-	public void registerItems(RegistryEvent.Register<Item> event) {
-		event.getRegistry().registerAll(elements.getItems().stream().map(Supplier::get).toArray(Item[]::new));
-	}
-
-	@SubscribeEvent
-	public void registerEntities(RegistryEvent.Register<EntityType<?>> event) {
-		event.getRegistry().registerAll(elements.getEntities().stream().map(Supplier::get).toArray(EntityType[]::new));
-	}
-
-	@SubscribeEvent
-	public void registerEnchantments(RegistryEvent.Register<Enchantment> event) {
-		event.getRegistry().registerAll(elements.getEnchantments().stream().map(Supplier::get).toArray(Enchantment[]::new));
-	}
-
-	@SubscribeEvent
-	public void registerSounds(RegistryEvent.Register<net.minecraft.util.SoundEvent> event) {
-		elements.registerSounds(event);
-	}
-
-	private static class FundysCursedFoodModFMLBusEvents {
-		private final FundysCursedFoodMod parent;
-
-		FundysCursedFoodModFMLBusEvents(FundysCursedFoodMod parent) {
-			this.parent = parent;
-		}
-
-		@SubscribeEvent
-		public void serverLoad(FMLServerStartingEvent event) {
-			this.parent.elements.getElements().forEach(element -> element.serverLoad(event));
-		}
+	public static <T> void addNetworkMessage(Class<T> messageType, BiConsumer<T, FriendlyByteBuf> encoder, Function<FriendlyByteBuf, T> decoder,
+			BiConsumer<T, Supplier<NetworkEvent.Context>> messageConsumer) {
+		PACKET_HANDLER.registerMessage(messageID, messageType, encoder, decoder, messageConsumer);
+		messageID++;
 	}
 }
