@@ -1,68 +1,39 @@
 
 package cz.pisekpiskovec.fundyscursedfood.item;
 
-import net.minecraftforge.registries.ObjectHolder;
 import net.minecraftforge.api.distmarker.OnlyIn;
 import net.minecraftforge.api.distmarker.Dist;
 
-import net.minecraft.world.World;
-import net.minecraft.util.text.StringTextComponent;
-import net.minecraft.util.text.ITextComponent;
-import net.minecraft.item.Rarity;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Item;
-import net.minecraft.item.Food;
-import net.minecraft.client.util.ITooltipFlag;
-import net.minecraft.block.BlockState;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.item.TooltipFlag;
+import net.minecraft.world.item.Rarity;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.food.FoodProperties;
+import net.minecraft.network.chat.TextComponent;
+import net.minecraft.network.chat.Component;
 
 import java.util.List;
 
-import cz.pisekpiskovec.fundyscursedfood.itemgroup.CreativeTabItemGroup;
-import cz.pisekpiskovec.fundyscursedfood.FundysCursedFoodModElements;
+import cz.pisekpiskovec.fundyscursedfood.init.FundysCursedFoodModTabs;
 
-@FundysCursedFoodModElements.ModElement.Tag
-public class SalmuttotabreabbitlecochicarrochopiePieItem extends FundysCursedFoodModElements.ModElement {
-	@ObjectHolder("fundys_cursed_food:salmuttotabreabbitlecochicarrochopie_pie")
-	public static final Item block = null;
+public class SalmuttotabreabbitlecochicarrochopiePieItem extends Item {
+	public SalmuttotabreabbitlecochicarrochopiePieItem() {
+		super(new Item.Properties().tab(FundysCursedFoodModTabs.TAB_CREATIVE_TAB).stacksTo(64).rarity(Rarity.COMMON)
+				.food((new FoodProperties.Builder()).nutrition(68).saturationMod(72f)
 
-	public SalmuttotabreabbitlecochicarrochopiePieItem(FundysCursedFoodModElements instance) {
-		super(instance, 12);
+						.meat().build()));
 	}
 
 	@Override
-	public void initElements() {
-		elements.items.add(() -> new ItemCustom());
+	@OnlyIn(Dist.CLIENT)
+	public boolean isFoil(ItemStack itemstack) {
+		return true;
 	}
 
-	public static class ItemCustom extends Item {
-		public ItemCustom() {
-			super(new Item.Properties().group(CreativeTabItemGroup.tab).maxStackSize(64).rarity(Rarity.COMMON)
-					.food((new Food.Builder()).hunger(68).saturation(72f)
-
-							.meat().build()));
-			setRegistryName("salmuttotabreabbitlecochicarrochopie_pie");
-		}
-
-		@Override
-		public int getItemEnchantability() {
-			return 0;
-		}
-
-		@Override
-		public float getDestroySpeed(ItemStack par1ItemStack, BlockState par2Block) {
-			return 1F;
-		}
-
-		@Override
-		@OnlyIn(Dist.CLIENT)
-		public boolean hasEffect(ItemStack itemstack) {
-			return true;
-		}
-
-		@Override
-		public void addInformation(ItemStack itemstack, World world, List<ITextComponent> list, ITooltipFlag flag) {
-			super.addInformation(itemstack, world, list, flag);
-			list.add(new StringTextComponent("\u00A7eit's glowing."));
-		}
+	@Override
+	public void appendHoverText(ItemStack itemstack, Level world, List<Component> list, TooltipFlag flag) {
+		super.appendHoverText(itemstack, world, list, flag);
+		list.add(new TextComponent("\u00A7eit's glowing."));
 	}
 }
