@@ -1,86 +1,52 @@
 
 package cz.pisekpiskovec.fundyscursedfood.item;
 
-import net.minecraftforge.registries.ObjectHolder;
 import net.minecraftforge.api.distmarker.OnlyIn;
 import net.minecraftforge.api.distmarker.Dist;
 
-import net.minecraft.world.World;
-import net.minecraft.util.text.StringTextComponent;
-import net.minecraft.util.text.ITextComponent;
-import net.minecraft.item.Rarity;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Item;
-import net.minecraft.item.Food;
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.client.util.ITooltipFlag;
-import net.minecraft.block.BlockState;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.item.TooltipFlag;
+import net.minecraft.world.item.Rarity;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.food.FoodProperties;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.network.chat.TextComponent;
+import net.minecraft.network.chat.Component;
 
-import java.util.stream.Stream;
-import java.util.Map;
 import java.util.List;
-import java.util.HashMap;
-import java.util.AbstractMap;
 
 import cz.pisekpiskovec.fundyscursedfood.procedures.DieEatingProcedureProcedure;
-import cz.pisekpiskovec.fundyscursedfood.itemgroup.CreativeTabItemGroup;
-import cz.pisekpiskovec.fundyscursedfood.FundysCursedFoodModElements;
+import cz.pisekpiskovec.fundyscursedfood.init.FundysCursedFoodModTabs;
 
-@FundysCursedFoodModElements.ModElement.Tag
-public class SalmuttotabreabbitlecochicarrochopotatiePieCakeItem extends FundysCursedFoodModElements.ModElement {
-	@ObjectHolder("fundys_cursed_food:salmuttotabreabbitlecochicarrochopotatie_pie_cake")
-	public static final Item block = null;
+public class SalmuttotabreabbitlecochicarrochopotatiePieCakeItem extends Item {
+	public SalmuttotabreabbitlecochicarrochopotatiePieCakeItem() {
+		super(new Item.Properties().tab(FundysCursedFoodModTabs.TAB_CREATIVE_TAB).stacksTo(64).rarity(Rarity.COMMON)
+				.food((new FoodProperties.Builder()).nutrition(81).saturationMod(79.60000000000001f)
 
-	public SalmuttotabreabbitlecochicarrochopotatiePieCakeItem(FundysCursedFoodModElements instance) {
-		super(instance, 14);
+						.meat().build()));
 	}
 
 	@Override
-	public void initElements() {
-		elements.items.add(() -> new ItemCustom());
+	@OnlyIn(Dist.CLIENT)
+	public boolean isFoil(ItemStack itemstack) {
+		return true;
 	}
 
-	public static class ItemCustom extends Item {
-		public ItemCustom() {
-			super(new Item.Properties().group(CreativeTabItemGroup.tab).maxStackSize(64).rarity(Rarity.COMMON)
-					.food((new Food.Builder()).hunger(81).saturation(79.60000000000001f)
+	@Override
+	public void appendHoverText(ItemStack itemstack, Level world, List<Component> list, TooltipFlag flag) {
+		super.appendHoverText(itemstack, world, list, flag);
+		list.add(new TextComponent("\u00A74this is bound to kill anyone who eats it."));
+	}
 
-							.meat().build()));
-			setRegistryName("salmuttotabreabbitlecochicarrochopotatie_pie_cake");
-		}
+	@Override
+	public ItemStack finishUsingItem(ItemStack itemstack, Level world, LivingEntity entity) {
+		ItemStack retval = super.finishUsingItem(itemstack, world, entity);
+		double x = entity.getX();
+		double y = entity.getY();
+		double z = entity.getZ();
 
-		@Override
-		public int getItemEnchantability() {
-			return 0;
-		}
-
-		@Override
-		public float getDestroySpeed(ItemStack par1ItemStack, BlockState par2Block) {
-			return 1F;
-		}
-
-		@Override
-		@OnlyIn(Dist.CLIENT)
-		public boolean hasEffect(ItemStack itemstack) {
-			return true;
-		}
-
-		@Override
-		public void addInformation(ItemStack itemstack, World world, List<ITextComponent> list, ITooltipFlag flag) {
-			super.addInformation(itemstack, world, list, flag);
-			list.add(new StringTextComponent("\u00A74this is bound to kill anyone who eats it."));
-		}
-
-		@Override
-		public ItemStack onItemUseFinish(ItemStack itemstack, World world, LivingEntity entity) {
-			ItemStack retval = super.onItemUseFinish(itemstack, world, entity);
-			double x = entity.getPosX();
-			double y = entity.getPosY();
-			double z = entity.getPosZ();
-
-			DieEatingProcedureProcedure.executeProcedure(Stream.of(new AbstractMap.SimpleEntry<>("entity", entity)).collect(HashMap::new,
-					(_m, _e) -> _m.put(_e.getKey(), _e.getValue()), Map::putAll));
-			return retval;
-		}
+		DieEatingProcedureProcedure.execute(entity);
+		return retval;
 	}
 }
