@@ -1,48 +1,17 @@
 
 package cz.pisekpiskovec.fundyscursedfood.item;
 
-import net.minecraftforge.registries.ObjectHolder;
+import net.minecraft.world.item.Rarity;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.food.FoodProperties;
 
-import net.minecraft.item.Rarity;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Item;
-import net.minecraft.item.Food;
-import net.minecraft.block.BlockState;
+import cz.pisekpiskovec.fundyscursedfood.init.FundysCursedFoodModTabs;
 
-import cz.pisekpiskovec.fundyscursedfood.itemgroup.CreativeTabItemGroup;
-import cz.pisekpiskovec.fundyscursedfood.FundysCursedFoodModElements;
+public class SteakchopItem extends Item {
+	public SteakchopItem() {
+		super(new Item.Properties().tab(FundysCursedFoodModTabs.TAB_CREATIVE_TAB).stacksTo(64).rarity(Rarity.COMMON)
+				.food((new FoodProperties.Builder()).nutrition(16).saturationMod(25.599999999999998f)
 
-@FundysCursedFoodModElements.ModElement.Tag
-public class SteakchopItem extends FundysCursedFoodModElements.ModElement {
-	@ObjectHolder("fundys_cursed_food:steakchop")
-	public static final Item block = null;
-
-	public SteakchopItem(FundysCursedFoodModElements instance) {
-		super(instance, 2);
-	}
-
-	@Override
-	public void initElements() {
-		elements.items.add(() -> new ItemCustom());
-	}
-
-	public static class ItemCustom extends Item {
-		public ItemCustom() {
-			super(new Item.Properties().group(CreativeTabItemGroup.tab).maxStackSize(64).rarity(Rarity.COMMON)
-					.food((new Food.Builder()).hunger(16).saturation(25.599999999999998f)
-
-							.meat().build()));
-			setRegistryName("steakchop");
-		}
-
-		@Override
-		public int getItemEnchantability() {
-			return 0;
-		}
-
-		@Override
-		public float getDestroySpeed(ItemStack par1ItemStack, BlockState par2Block) {
-			return 1F;
-		}
+						.meat().build()));
 	}
 }
